pyproject.toml
[tool.poetry]
name = "fastapi-template"
version = "0.1.0"
description = ""
authors = ["rlaminseok0824 <xornjsrlaals@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.111.1"
uvicorn = {extras = ["standard"], version = "^0.30.4"}
python-multipart = "^0.0.9"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pydantic = "^2.8.2"


alembic = "^1.13.2"
pydantic-settings = "^2.4.0"
asyncpg = "^0.29.0"
psycopg = {extras = ["binary"], version = "^3.2.1"}
greenlet = "^3.0.3"
pyjwt = "^2.9.0"
bcrypt = "^4.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
